+++ swift-beacon-chain/Sources/DataStructures/State/DepositRootVote.swift
@@ -10,5 +10,5 @@ import Foundation
 
 struct DepositRootVote {
     let depositRoot: Data;
-    let voteCount: uint64;
+    let voteCount: Int;
 }
